now k-s can get reasonable "smoothed" nl pp. Let's test how does it work

+++ interface/vps_params/proto_vps_params_local.h
@@ -60,7 +60,7 @@ void make_weight_gen(double *,double *,double *, double *,
 void nlppSmoothKS(PSEUDO *,double *,double ,int ,double *,int ,double ,int ,int);
 void nlppSmoothRoi(PSEUDO *,double *,double ,int ,double *,int ,double ,int ,int);
 void bessTransform(double *,int ,double ,int ,double *,int ,double );
-void optGCoeff(PSEUDO_REAL *,int ,int ,int ,double ,double ,double ,int ,double *);
+void optGCoeff(PSEUDO_REAL *,int ,int ,int ,double ,double ,int ,int ,double *);
